journal details refactor
